pyproject.toml
[project]
name = "ir-project-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.0",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.3",
    "pandas>=2.3.3",
    "pdfminer>=20191125",
    "pydantic>=2.12.3",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
